pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "rembg>=2.0.67",
    "tqdm>=4.67.1",
]
